requirements.txt
futures3
mysql-connector-python
requests-futures
